+++ src/client/introspection.rs
@@ -261,7 +261,7 @@ mod tests {
         E2E_DEVICE_PROPERTY_NAME,
     };
     use crate::validate::ValidatedIndividual;
-    use crate::AstarteType;
+    use crate::AstarteData;
 
     #[tokio::test]
     async fn get_interface() {
@@ -400,7 +400,7 @@ mod tests {
                     version_major: 0,
                     reliability: Reliability::Guaranteed,
                     retention: Retention::Volatile { expiry: None },
-                    data: AstarteType::try_from(42.0).unwrap(),
+                    data: AstarteData::try_from(42.0).unwrap(),
                     timestamp: Some(Utc::now()),
                 },
             )
@@ -453,7 +453,7 @@ mod tests {
                     version_major: 0,
                     reliability: Reliability::Guaranteed,
                     retention: Retention::Stored { expiry: None },
-                    data: AstarteType::try_from(42.0).unwrap(),
+                    data: AstarteData::try_from(42.0).unwrap(),
                     timestamp: Some(Utc::now()),
                 },
                 &[1, 2, 3, 4],
@@ -509,7 +509,7 @@ mod tests {
                     version_major: 0,
                     reliability: Reliability::Guaranteed,
                     retention: Retention::Volatile { expiry: None },
-                    data: AstarteType::try_from(42.0).unwrap(),
+                    data: AstarteData::try_from(42.0).unwrap(),
                     timestamp: Some(Utc::now()),
                 },
             )
@@ -565,7 +565,7 @@ mod tests {
                     version_major: 0,
                     reliability: Reliability::Guaranteed,
                     retention: Retention::Stored { expiry: None },
-                    data: AstarteType::try_from(42.0).unwrap(),
+                    data: AstarteData::try_from(42.0).unwrap(),
                     timestamp: Some(Utc::now()),
                 },
                 &[1, 2, 3, 4],
@@ -662,7 +662,7 @@ mod tests {
             .store_prop(crate::store::StoredProp {
                 interface: E2E_DEVICE_PROPERTY_NAME,
                 path: "/sensor_1/double_endpoint",
-                value: &AstarteType::LongInteger(2),
+                value: &AstarteData::LongInteger(2),
                 interface_major: to_remove.version_major(),
                 ownership: to_remove.ownership(),
             })