+++ src/client/object.rs
@@ -202,7 +202,7 @@ mod tests {
         E2E_DEVICE_DATASTREAM, E2E_DEVICE_DATASTREAM_NAME, STORED_DEVICE_OBJECT,
         STORED_DEVICE_OBJECT_NAME, VOLATILE_DEVICE_OBJECT, VOLATILE_DEVICE_OBJECT_NAME,
     };
-    use crate::{AstarteType, Client};
+    use crate::{AstarteData, Client};
 
     #[tokio::test]
     async fn send_datastream_object_connected_discard() {
@@ -214,12 +214,12 @@ mod tests {
 
         let obj = AstarteObject::from_iter(
             [
-                ("double_endpoint", AstarteType::try_from(42.0).unwrap()),
-                ("integer_endpoint", AstarteType::Integer(42)),
-                ("boolean_endpoint", AstarteType::Boolean(false)),
+                ("double_endpoint", AstarteData::try_from(42.0).unwrap()),
+                ("integer_endpoint", AstarteData::Integer(42)),
+                ("boolean_endpoint", AstarteData::Boolean(false)),
                 (
                     "booleanarray_endpoint",
-                    AstarteType::BooleanArray(vec![true, false]),
+                    AstarteData::BooleanArray(vec![true, false]),
                 ),
             ]
             .map(|(n, v)| (n.to_string(), v)),
@@ -260,8 +260,8 @@ mod tests {
         let path = "/endpoint";
         let value = AstarteObject::from_iter(
             [
-                ("longinteger", AstarteType::LongInteger(42)),
-                ("boolean", AstarteType::Boolean(true)),
+                ("longinteger", AstarteData::LongInteger(42)),
+                ("boolean", AstarteData::Boolean(true)),
             ]
             .map(|(k, v)| (k.to_string(), v)),
         );
@@ -309,8 +309,8 @@ mod tests {
         let path = "/endpoint";
         let value = AstarteObject::from_iter(
             [
-                ("longinteger", AstarteType::LongInteger(42)),
-                ("boolean", AstarteType::Boolean(true)),
+                ("longinteger", AstarteData::LongInteger(42)),
+                ("boolean", AstarteData::Boolean(true)),
             ]
             .map(|(k, v)| (k.to_string(), v)),
         );
@@ -360,8 +360,8 @@ mod tests {
         let path = "/endpoint";
         let value = AstarteObject::from_iter(
             [
-                ("longinteger", AstarteType::LongInteger(42)),
-                ("boolean", AstarteType::Boolean(true)),
+                ("longinteger", AstarteData::LongInteger(42)),
+                ("boolean", AstarteData::Boolean(true)),
             ]
             .map(|(k, v)| (k.to_string(), v)),
         );
@@ -439,12 +439,12 @@ mod tests {
 
         let obj = AstarteObject::from_iter(
             [
-                ("double_endpoint", AstarteType::try_from(42.0).unwrap()),
-                ("integer_endpoint", AstarteType::Integer(42)),
-                ("boolean_endpoint", AstarteType::Boolean(false)),
+                ("double_endpoint", AstarteData::try_from(42.0).unwrap()),
+                ("integer_endpoint", AstarteData::Integer(42)),
+                ("boolean_endpoint", AstarteData::Boolean(false)),
                 (
                     "booleanarray_endpoint",
-                    AstarteType::BooleanArray(vec![true, false]),
+                    AstarteData::BooleanArray(vec![true, false]),
                 ),
             ]
             .map(|(n, v)| (n.to_string(), v)),
@@ -466,8 +466,8 @@ mod tests {
         let path = "/endpoint";
         let value = AstarteObject::from_iter(
             [
-                ("longinteger", AstarteType::LongInteger(42)),
-                ("boolean", AstarteType::Boolean(true)),
+                ("longinteger", AstarteData::LongInteger(42)),
+                ("boolean", AstarteData::Boolean(true)),
             ]
             .map(|(k, v)| (k.to_string(), v)),
         );
@@ -503,8 +503,8 @@ mod tests {
         let path = "/endpoint";
         let value = AstarteObject::from_iter(
             [
-                ("longinteger", AstarteType::LongInteger(42)),
-                ("boolean", AstarteType::Boolean(true)),
+                ("longinteger", AstarteData::LongInteger(42)),
+                ("boolean", AstarteData::Boolean(true)),
             ]
             .map(|(k, v)| (k.to_string(), v)),
         );
@@ -543,8 +543,8 @@ mod tests {
         let path = "/endpoint";
         let value = AstarteObject::from_iter(
             [
-                ("longinteger", AstarteType::LongInteger(42)),
-                ("boolean", AstarteType::Boolean(true)),
+                ("longinteger", AstarteData::LongInteger(42)),
+                ("boolean", AstarteData::Boolean(true)),
             ]
             .map(|(k, v)| (k.to_string(), v)),
         );
@@ -613,12 +613,12 @@ mod tests {
 
         let obj = AstarteObject::from_iter(
             [
-                ("double_endpoint", AstarteType::try_from(42.0).unwrap()),
-                ("integer_endpoint", AstarteType::Integer(42)),
-                ("boolean_endpoint", AstarteType::Boolean(false)),
+                ("double_endpoint", AstarteData::try_from(42.0).unwrap()),
+                ("integer_endpoint", AstarteData::Integer(42)),
+                ("boolean_endpoint", AstarteData::Boolean(false)),
                 (
                     "booleanarray_endpoint",
-                    AstarteType::BooleanArray(vec![true, false]),
+                    AstarteData::BooleanArray(vec![true, false]),
                 ),
             ]
             .map(|(n, v)| (n.to_string(), v)),
@@ -649,12 +649,12 @@ mod tests {
 
         let obj = AstarteObject::from_iter(
             [
-                ("double_endpoint", AstarteType::try_from(42.0).unwrap()),
-                ("integer_endpoint", AstarteType::Integer(42)),
-                ("boolean_endpoint", AstarteType::Boolean(false)),
+                ("double_endpoint", AstarteData::try_from(42.0).unwrap()),
+                ("integer_endpoint", AstarteData::Integer(42)),
+                ("boolean_endpoint", AstarteData::Boolean(false)),
                 (
                     "booleanarray_endpoint",
-                    AstarteType::BooleanArray(vec![true, false]),
+                    AstarteData::BooleanArray(vec![true, false]),
                 ),
             ]
             .map(|(n, v)| (n.to_string(), v)),