+++ src/client/property.rs
@@ -27,7 +27,7 @@ use crate::state::Status;
 use crate::store::{PropertyMapping, PropertyStore, StoredProp};
 use crate::transport::Connection;
 use crate::validate::{ValidatedProperty, ValidatedUnset};
-use crate::{AstarteType, Error};
+use crate::{AstarteData, Error};
 
 use super::{DeviceClient, MappingRef, Publish};
 
@@ -39,7 +39,7 @@ where
         &mut self,
         interface_name: &str,
         path: &MappingPath<'_>,
-        data: AstarteType,
+        data: AstarteData,
     ) -> Result<(), Error>
     where
         C::Sender: Publish,
@@ -109,7 +109,7 @@ where
         &self,
         mapping: &MappingRef<'_, PropertyRef<'_>>,
         path: &MappingPath<'_>,
-    ) -> Result<Option<AstarteType>, Error> {
+    ) -> Result<Option<AstarteData>, Error> {
         let interface = mapping.interface();
         let path = path.as_str();
         let interface_data = PropertyMapping::with_property_path(interface, path);
@@ -185,7 +185,7 @@ mod tests {
     use crate::store::{PropertyInterface, PropertyMapping, PropertyStore, StoredProp};
     use crate::test::{E2E_DEVICE_PROPERTY, E2E_DEVICE_PROPERTY_NAME};
     use crate::validate::{ValidatedProperty, ValidatedUnset};
-    use crate::{AstarteType, Client};
+    use crate::{AstarteData, Client};
 
     #[tokio::test]
     async fn send_property_connected() {
@@ -194,7 +194,7 @@ mod tests {
         client.state.status.set_connected(true);
 
         let path = "/sensor_1/longinteger_endpoint";
-        let value = AstarteType::LongInteger(42);
+        let value = AstarteData::LongInteger(42);
 
         let mut seq = Sequence::new();
 
@@ -239,7 +239,7 @@ mod tests {
         client.state.status.set_connected(false);
 
         let path = "/sensor_1/longinteger_endpoint";
-        let value = AstarteType::LongInteger(42);
+        let value = AstarteData::LongInteger(42);
 
         // Send
         client
@@ -269,7 +269,7 @@ mod tests {
         client.state.status.set_connected(true);
 
         let path = "/sensor_1/longinteger_endpoint";
-        let value = AstarteType::LongInteger(42);
+        let value = AstarteData::LongInteger(42);
 
         // No expect, but store the prop
         client
@@ -354,7 +354,7 @@ mod tests {
         client.state.status.set_connected(true);
 
         let path = "/sensor_1/longinteger_endpoint";
-        let value = AstarteType::LongInteger(42);
+        let value = AstarteData::LongInteger(42);
 
         client
             .store
@@ -362,7 +362,7 @@ mod tests {
                 interface: E2E_DEVICE_PROPERTY_NAME,
                 path,
                 // Wrong type
-                value: &AstarteType::Boolean(false),
+                value: &AstarteData::Boolean(false),
                 interface_major: 0,
                 ownership: Ownership::Device,
             })